chore: add routes
